replace packager.Pull with packager2.Pull impl and add options struct
